FPAD-8430: Separate status retriever handler lambda and logic into dedicated files
